+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+#### HEAD
+
+- Added an `isISRC()` validator for [ISRC](https://en.wikipedia.org/wiki/International_Standard_Recording_Code)
+  ([#660](https://github.com/chriso/validator.js/pull/660))
+- New locales
+  ([#647](https://github.com/chriso/validator.js/pull/647))
+
 #### 7.0.0
 
 - Remove `isDate()`